requirements.txt
RPi.GPIO>=0.7.0,<1.0.0
hivemind_voice_relay
ovos-bus-client